refactor: extract utilities and remove dead code from   App.tsx
